Created "SpawnGrass" script
Created a script for spawning objects around the map outside of the player's view and at least a certain distance away from any other object with a collider. We just need to plug some values in to the script/object when the actual map is created, see comments on the "SpawnGrass" script for details.